requirements.txt
langchain
langchain-community
langgraph
sqlalchemy
mysql-connector-python
openai
python-dotenv 